refactor: remove DB dispatch from canvas component
